installer/dev/installer/test/com/distocraft/dc5000/install/ant/ParseChangelogTest.java
package com.distocraft.dc5000.install.ant;

import static org.junit.Assert.*;

import java.io.File;
import java.io.FileOutputStream;
import java.io.PrintWriter;
import java.io.FileNotFoundException;
import java.lang.reflect.Field;
import java.lang.reflect.Method;
import java.sql.Connection;
import java.sql.DriverManager;
import java.sql.SQLException;
import java.sql.Statement;
import java.sql.Timestamp;
import java.util.HashMap;
import java.util.Iterator;
import java.util.Vector;
import java.net.URL;

import junit.framework.JUnit4TestAdapter;

import org.dbunit.Assertion;
import org.dbunit.database.DatabaseConnection;
import org.dbunit.dataset.IDataSet;
import org.dbunit.dataset.ITable;
import org.dbunit.dataset.xml.FlatXmlDataSet;
import org.dbunit.dataset.xml.FlatXmlWriter;
import org.junit.AfterClass;
import org.junit.Ignore;
import org.junit.Test;
import org.junit.BeforeClass;

import com.distocraft.dc5000.install.ant.ParseChangelog.EntryDetails;

import ssc.rockfactory.RockFactory;

/**
 * 
 * @author ejarsok
 *
 */

public class ParseChangelogTest {

  private static Method insertEntryDetailsToDatabase;

  private static Method getPreviousTagBuildDate;

  private static Method checkForDefectEntry;

  private static Field rockFactory;

  private static Field entryDetailsMap;

  private static RockFactory rockFact;
  
  private static Connection c;
  
  private static Statement stm;

  @BeforeClass
  public static void init() {

    try {
      Class.forName("org.hsqldb.jdbcDriver");
    } catch (ClassNotFoundException e2) {
      e2.printStackTrace();
      fail("init() failed, ClassNotFoundException");
    }
    
    try {
      c = DriverManager.getConnection("jdbc:hsqldb:mem:testdb", "SA", "");
      stm = c.createStatement();

      stm
          .execute("CREATE TABLE Build (BUILDNUMBER BIGINT, MODULE VARCHAR(31), AUTHOR VARCHAR(31), BUILDDATE TIMESTAMP, "
              + "BUILDTAG VARCHAR(31), MODULETESTER VARCHAR(31), MODULETESTDATE TIMESTAMP, TESTRESULT VARCHAR(31))");

      stm
          .executeUpdate("INSERT INTO Build VALUES(10, 'module', 'emickmous', '2008-07-21 00:00:00.0', 'tag', 'tester', 2008-07-21, 'result')");
      stm
          .executeUpdate("INSERT INTO Build VALUES(10, 'module', 'emickmous', '2008-07-20 00:00:00.0', 'tag', 'tester', 2008-07-20, 'result')");
      
   
      stm.execute("CREATE TABLE Moduledefect (BUILDNUMBER BIGINT, MODULE VARCHAR(31), TRACKERPROJECT VARCHAR(31), "
          + "TRACKERID VARCHAR(31))");

      stm
          .execute("CREATE TABLE Moduleinfo (BUILDNUMBER BIGINT, MODULE VARCHAR(31), TYPE VARCHAR(31), SUMMARY VARCHAR(31), "
              + "DESCRIPTION VARCHAR(31))");

    } catch (SQLException e1) {
      e1.printStackTrace();
      fail("init() failed, SQLException");
    }

    ParseChangelog pcl = new ParseChangelog();
    Class secretClass = pcl.getClass();
    try {
      rockFact = new RockFactory("jdbc:hsqldb:mem:testdb", "SA", "", "org.hsqldb.jdbcDriver", "con", true, -1);

      insertEntryDetailsToDatabase = secretClass.getDeclaredMethod("insertEntryDetailsToDatabase", null);
      getPreviousTagBuildDate = secretClass.getDeclaredMethod("getPreviousTagBuildDate", null);
      checkForDefectEntry = secretClass.getDeclaredMethod("checkForDefectEntry", EntryDetails.class);
      rockFactory = secretClass.getDeclaredField("rockFactory");
      entryDetailsMap = secretClass.getDeclaredField("entryDetailsMap");
      insertEntryDetailsToDatabase.setAccessible(true);
      checkForDefectEntry.setAccessible(true);
      getPreviousTagBuildDate.setAccessible(true);
      rockFactory.setAccessible(true);
      entryDetailsMap.setAccessible(true);
    } catch (Exception e) {
      e.printStackTrace();
      fail("init() failed");
    }
  }
  
  @AfterClass
  public static void tearDownAfterClass() throws Exception {
    stm.execute("DROP TABLE Build");
    stm.execute("DROP TABLE Moduledefect");
    stm.execute("DROP TABLE Moduleinfo");
  }

  @Test
  public void testSetAndGetBuildNumber() {
    ParseChangelog pcl = new ParseChangelog();
    pcl.setBuildNumber("b10");
    assertEquals("b10", pcl.getBuildNumber());
  }

  @Test
  public void testSetAndGetChangelogPath() {
    ParseChangelog pcl = new ParseChangelog();
    pcl.setChangelogPath("PATH");
    assertEquals("PATH", pcl.getChangelogPath());
  }

  @Test
  public void testSetAndGetModule() {
    ParseChangelog pcl = new ParseChangelog();
    pcl.setModule("module");
    assertEquals("module", pcl.getModule());
  }

  @Test
  public void testSetAndGetTag() {
    ParseChangelog pcl = new ParseChangelog();
    pcl.setTag("tag");
    assertEquals("tag", pcl.getTag());
  }

  @Test
  public void testInitializeRock() {
    ParseChangelog pcl = new ParseChangelog();
    try {
      RockFactory rockFact = null;
      rockFact = pcl.initializeRock();

      assertNotNull(rockFact);

    } catch (Exception e) {
      e.printStackTrace();
      fail("testInitializeRock() failed");
    }
  }

  @Test
  public void testGetPreviousTagBuildDate() {
    ParseChangelog pcl = new ParseChangelog();
    pcl.setModule("module");
    Timestamp t = new Timestamp(2008 - 1900, 6, 20, 0, 0, 0, 0);

    try {
      rockFactory.set(pcl, rockFact);
      Long l = new Long(0);

      l = (Long) getPreviousTagBuildDate.invoke(pcl, null);
      assertEquals((Long) t.getTime(), l);

    } catch (Exception e) {
      e.printStackTrace();
      fail("testGetPreviousTagBuildDate() failed");
    }
  }

  @Test
  public void testCheckForDefectEntry() {
    ParseChangelog pcl = new ParseChangelog();
    EntryDetails ed = pcl.new EntryDetails();
    ed.message = "message";
    try {
      assertEquals(false, checkForDefectEntry.invoke(pcl, ed));

    } catch (Exception e) {
      e.printStackTrace();
      fail("testCheckForDefectEntry() failed");
    }
  }

  @Test
  public void testCheckForDefectEntry2() {
    ParseChangelog pcl = new ParseChangelog();
    pcl.setBuildNumber("1");
    pcl.setModule("modul");
    EntryDetails ed = pcl.new EntryDetails();
    ed.message = "[Defect#123@Platform_rel5]";
    try {
      rockFactory.set(pcl, rockFact);
      assertEquals(true, checkForDefectEntry.invoke(pcl, ed));

    } catch (Exception e) {
      e.printStackTrace();
      fail("testCheckForDefectEntry2() failed");
    }
  }

  @Ignore
  public void testInsertEntryDetailsToDatabase() {
    ParseChangelog pcl = new ParseChangelog();
    pcl.setBuildNumber("10");
    pcl.setModule("moduleName");
    Vector v = new Vector();
    HashMap hm = new HashMap();
    HashMap hm2 = new HashMap();

    EntryDetails ed1 = pcl.new EntryDetails();
    ed1.date = "2008-07-21";
    ed1.time = "00:00";
    ed1.message = "ed1_message";

    EntryDetails ed2 = pcl.new EntryDetails();
    ed2.date = "2008-07-21";
    ed2.time = "00:00";
    ed2.message = "ed2_message";

    hm.put("filename1", "1.1.1");
    ed1.fileChanges = hm;

    hm2.put("filename1", "1.1.2");
    ed2.fileChanges = hm2;

    v.add(ed1);
    v.add(ed2);

    try {
      rockFactory.set(pcl, rockFact);
      entryDetailsMap.set(pcl, v);

      insertEntryDetailsToDatabase.invoke(pcl, null);

      
      IDataSet actualDataSet = new DatabaseConnection(c).createDataSet();
      ITable actualTable = actualDataSet.getTable("Moduleinfo");
      //ITable actualTable = new DatabaseConnection(c).createQueryTable("RESULT_NAME",
      //    "SELECT * FROM Moduleinfo");
      IDataSet expectedDataSet = new FlatXmlDataSet(getFile(
          "com.distocraft.dc5000.install.ant_ParseChangelog_testInsertEntryDetailsToDatabase/Expected.xml"));
      ITable expectedTable = expectedDataSet.getTable("Moduleinfo");

      File f = new File(System.getProperty("user.home"), "testi.xml");
      FlatXmlWriter fw = new FlatXmlWriter(new FileOutputStream(f));
      fw.write(actualDataSet);
      
      
      System.out.println("---------------------------------------");
      System.out.println(actualTable.getValue(0, "DESCRIPTION"));
      System.out.println("---------------------------------------");
      System.out.println(expectedTable.getValue(0, "DESCRIPTION"));
      System.out.println("---------------------------------------");

      // TODO Test and data is ok but assertion says that expectedTable != actualTable
      Assertion.assertEquals(expectedTable, actualTable);
      
    } catch (Exception e) {
      e.printStackTrace();
      fail("testInsertEntryDetailsToDatabase() failed");
    }
  }
	private File getFile(final String name) throws Exception {
		final URL url = ClassLoader.getSystemResource("XMLFiles");
		if(url == null){
			throw new FileNotFoundException("XMLFiles");
		}
		final File xmlBase = new File(url.toURI());
		final String xmlFile = xmlBase.getAbsolutePath() + "/"+name;
		return new File(xmlFile);
	}

  @Ignore
  public void testExecute() {
    // TODO RockFactory problem
    Field entryDetailsMap = null;
    
    ParseChangelog pcl = new ParseChangelog();
    File xmlFile = new File(System.getProperty("user.home"), "xmlFile.xml");
    pcl.setChangelogPath(xmlFile.getAbsolutePath());
    pcl.setTag("tag");
    pcl.setModule("moduleName");
    pcl.setBuildNumber("20");
    
    Class secretClass = pcl.getClass();
    
    try {
      entryDetailsMap = secretClass.getDeclaredField("entryDetailsMap");
      
      entryDetailsMap.setAccessible(true);
    } catch (Exception e1) {
      e1.printStackTrace();
      fail("testExecute() failed");
    }
    
    try{
      PrintWriter pw = new PrintWriter(new FileOutputStream(xmlFile));
      pw.write("<?xml version=\"1.0\" encoding=\"UTF-8\" ?>\n");
      pw.write("<dataset>\n");
      pw.write("<entry>\n");
      pw.write("<date>2008-07-21</date>\n");
      pw.write("<time>00:00</time>\n");
      pw.write("<msg>Message</msg>\n");
      pw.write("<file>\n");
      pw.write("<name>fileName_Foobar</name>\n");
      pw.write("<revision>1.1.2</revision>\n");
      pw.write("</file>\n");
      pw.write("</entry>\n");
      pw.write("</dataset>\n");
      pw.close();
    } catch(Exception e) {
      fail("testExecute() failed, can't write in file");
    }
      
    try {
      pcl.execute();
      
      IDataSet actualDataSet = new DatabaseConnection(c).createDataSet();
      ITable actualTable = actualDataSet.getTable("Moduleinfo");
    
      File f = new File(System.getProperty("user.home"), "xmlFile_actual.xml");
      FlatXmlWriter fw = new FlatXmlWriter(new FileOutputStream(f));
      fw.write(actualDataSet);
    } catch(Exception e) {
      e.printStackTrace();
      fail("testExecute() failed");
    }
  }

  /*public static junit.framework.Test suite() {
    return new JUnit4TestAdapter(ParseChangelogTest.class);
  }*/
}
